[Draft] change JAX type defaults to 32-bit
This is an experiment in (eventually) removing the jax_enable_x64 flag. Step 1 of the process outlined in https://github.com/google/jax/pull/8180

PiperOrigin-RevId: 405899983

+++ jax/experimental/jax2tf/tests/jax2tf_test.py
@@ -117,9 +117,9 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     mk_sharded = lambda f: jax.pmap(lambda x: x)(f([n]))
     f_tf = tf.function(lambda x: x)
     self.assertAllClose(f_tf(mk_sharded(jnp.zeros)).numpy(),
-                        np.zeros([n]))
+                        np.zeros([n], dtype=jnp.float_))
     self.assertAllClose(f_tf(mk_sharded(jnp.ones)).numpy(),
-                        np.ones([n]))
+                        np.ones([n], dtype=jnp.float_))
 
   @jtu.skip_on_devices("gpu")
   def test_bfloat16_passed_by_tf(self):
